feat: add tm-location-selector web component
This adds the tm-location-selector web component that should work out of
the box when used in plain HTML.

The following example shows a map, centered of Berlin.
After 10 seconds the location switches to somewhere in Sweden.
Users can select a location by clicking/tapping on the map.
Consumers can listen for the change event.

```html
<tm-location-selector
    id="location-selector"
    style="width: 500px; height: 300px; max-width: 100%;"
    label="Dein häufig besuchter Ort"
    lat="52.52"
    lng="13.4050"
></tm-location-selector>

<script type="module">
    const el = document.getElementById('location-selector')

    el.addEventListener('change', (event) => {
        const { lat, lng } = event.detail[0]
        console.log(`new position: ${lat}, ${lng}`)
    })

    setTimeout(() => {
        el.lat = 60.293
        el.label = 'Somewhere over in Sweden'
    }, 10000)
</script>
```

refs #19